+++ src/Phan/Bootstrap.php
@@ -18,6 +18,8 @@ if (extension_loaded('ast')) {
     // Warn if the php-ast version is too low.
     $ast_version = (string)phpversion('ast');
     if (version_compare($ast_version, '1.0.0') <= 0) {
+        // TODO: Change this to a warning for 0.1.5 - 1.0.0. (https://github.com/phan/phan/issues/2954)
+        // 0.1.5 introduced the ast\Node constructor, which is required by the polyfill
         fprintf(
             STDERR,
             "ERROR: Phan 2.x requires php-ast 1.0.1+ because it depends on AST version 70. php-ast %s is installed." . PHP_EOL,